feature/#20: Padding이 들어간 UILabel을 공통 컴포넌트로 분리
